add json_ops utils mod
