+++ packages/cw-jsonfilter/src/json_ops/mod.rs
@@ -5,3 +5,4 @@ pub mod op_logical;
 pub mod op_transform;
 pub mod op_type;
 pub mod operator;
+pub mod utils;
