+++ packages/cw-jsonfilter/src/json_ops/op_logical.rs
@@ -1,8 +1,6 @@
-// TODO: and, or, xor, not
-
 use crate::{
-    filter::append_array_path, json_ops::operator::OperatorContext, CwJsonFilter, FilterResult,
-    ProtobufDecoder,
+    json_ops::{operator::OperatorContext, utils::append_array_path},
+    CwJsonFilter, FilterResult, ProtobufDecoder,
 };
 
 impl<D: ProtobufDecoder> CwJsonFilter<D> {
